Trait galvanic_assert::Matcher

pub trait Matcher<'a, T: 'a> {
    fn check(&self, actual: &'a T) -> MatchResult;
}

The trait which has to be implemented by all matchers.

Required methods

fn check(&self, actual: &'a T) -> MatchResult

Checks the passed value if it satisfies the Matcher.

Values are always taken as immutable reference as the actual value shouldn't be changed by the matcher.

Loading content...